All Rooms Designed Stylistically
Ideally, I will come back to this and refactor it in general, but a few things in particular.

Figure out how to update clickable placement, only the backgrounds are moving on screen resize.

Also implement one for vertical resize.  My thoughts on this are that if people have smaller screens, it might appear janky.

Also, just a general refactor of how I'm loading the images/objects.  It seems like there should be a better way, especially since they are position absolute.