Top Bar Started
It may very well be finished.  I may deviate from the plan, and if so, it is finished.

Upon testing it out, I do notice that the character sheet isn't rendering properly on transition, as the 'Begin Quest' button should say 'Return' or the like, and the 'Choose Character' button should be gone after character creation.