Figured Out Media Size Point and Click I believe
It doesn't look super pretty, but I left the absolute positioning and instead of having everything in pixels (width/height/x/y) I have it at percentages.

It seems to be working so far, but took a decent bit of time to get it figured out and I am only done with Start.